64-bit iOS.
[FIXED]     More typing and format strings for platform-dependant types.
[FIXED]     Disabled iOS-only err strings on OSX.
[ADDED]     iOS 64-bit arch added to scryptenc and opensslcrypto libs.
[FIXED]     Added back missing header files to .m files.

+++ Pearl-Cocos2D/View/PearlCCBarSprite.h
@@ -32,7 +32,7 @@
     CGPoint _target;
 
     CGPoint _current;
-    CGFloat _currentLength;
+    float   _currentLength;
 
     CGSize     _textureSize;
     NSUInteger _uniformColor;
